Support Foundry

+++ contracts/IBallot.sol
@@ -9,16 +9,14 @@ interface IBallot {
     error CommitmentDeadlineNotPassed(uint256 commitmentDeadline);
     error InvalidOption(uint256 option);
 
-    function registerCommitment(
-        uint256 commitment
-    ) external;
+    function registerCommitment(uint256 commitment) external;
 
     function vote(
         uint256 option,
         uint256 nullifier,
         uint256 root,
-        uint[2] calldata proof_a,
-        uint[2][2] calldata proof_b,
-        uint[2] calldata proof_c
+        uint256[2] calldata proof_a,
+        uint256[2][2] calldata proof_b,
+        uint256[2] calldata proof_c
     ) external;
 }
